refactor: renamed path of api call functions from utils to services
