+++ src/components/VideoCard.js
@@ -3,7 +3,7 @@ import { useNavigate } from "react-router-dom";
 import { useAuth } from "../context/auth-context";
 import { useVideoAction } from "../context/video-action-context";
 import "../styles/components/videocard.css";
-import { addToHistory } from "../utils/addToHistory";
+import { addToHistory } from "../services/addToHistory";
 import { VideoOptions } from "./VideoOptions";
 
 export const VideoCard = ({ videoCardDetails }) => {
