+++ src/components/VideoOptions.js
@@ -2,30 +2,30 @@ import "../styles/components/videoOptions.css";
 import { useVideoAction } from "../context/video-action-context";
 import { RequiresAuth } from "../routes/RequiresAuth";
 import { useAuth } from "../context/auth-context";
-import { addToWatchLater } from "../utils/addToWatchLater";
-import { removeFromWatchLater } from "../utils/removeFromWatchLater";
-import { addToLikes } from "../utils/addToLikes";
-import { removeFromLikes } from "../utils/removeFromLikes";
-import { removeFromHistory } from "../utils/removeFromHistory";
+import { addToWatchLater } from "../services/addToWatchLater";
+import { removeFromWatchLater } from "../services/removeFromWatchLater";
+import { addToLikes } from "../services/addToLikes";
+import { removeFromLikes } from "../services/removeFromLikes";
+import { removeFromHistory } from "../services/removeFromHistory";
 
 export const VideoOptions = ({ activeVideo }) => {
   const { videoActionState, dispatchVideoAction } = useVideoAction();
   const { auth } = useAuth();
 
-  const isHistoryPage = () => {
+  const inHistoryPage = () => {
     let currentPathname = window.location.pathname;
     return currentPathname === "/history" ? true : false;
   };
 
-  const isInWatchLaterList = (videoId) => {
+  const inWatchLaterList = (videoId) => {
     return videoActionState.watchlater.some((item) => item._id === videoId);
   };
-  const isInLikesList = (videoId) => {
+  const inLikesList = (videoId) => {
     return videoActionState.likes.some((item) => item._id === videoId);
   };
 
   const watchLaterOnClickHandler = () => {
-    isInWatchLaterList(activeVideo._id)
+    inWatchLaterList(activeVideo._id)
       ? removeFromWatchLater({
           auth,
           activeVideo,
@@ -35,7 +35,7 @@ export const VideoOptions = ({ activeVideo }) => {
   };
 
   const likeVideoOnClickHandler = () => {
-    isInLikesList(activeVideo._id)
+    inLikesList(activeVideo._id)
       ? removeFromLikes({
           auth,
           activeVideo,
@@ -50,7 +50,7 @@ export const VideoOptions = ({ activeVideo }) => {
         <ul>
           <li onClick={watchLaterOnClickHandler}>
             <i className="material-icons">watch_later</i>
-            {isInWatchLaterList(activeVideo._id)
+            {inWatchLaterList(activeVideo._id)
               ? `Remove from watch later`
               : `Save to watch later`}
           </li>
@@ -68,11 +68,11 @@ export const VideoOptions = ({ activeVideo }) => {
           </li>
           <li onClick={likeVideoOnClickHandler}>
             <i className="material-icons">thumb_up</i>
-            {isInLikesList(activeVideo._id)
+            {inLikesList(activeVideo._id)
               ? `Remove from liked videos`
               : `Add to liked videos`}
           </li>
-          {isHistoryPage() && (
+          {inHistoryPage() && (
             <li
               onClick={() =>
                 removeFromHistory({
